remove funding metadata
Please contribute to the original tmux project.
The links result of the metadata from the original
project makes it appear like you're donating to tmux-rs
when you're donating to the tmux developer. If people
are able I'd like them to donate to the original tmux project.
But I don't want any confusion about who they may be
donating money to, so I'm removing the funding metadata
from this tmux-rs repo.